pi/requirements.txt
pyrebase4
opencv-python-headless